Move project list item delegate to new file
